requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
huggingface-hub>=0.19.0

# Quantization
optimum>=1.14.0
auto-gptq>=0.5.0
bitsandbytes>=0.41.0

# Memory management
pynvml>=11.5.0
psutil>=5.9.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
tensorboard>=2.15.0
wandb>=0.16.0

# API (for future use)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0